add: add library music feature

+++ node_server/src/controllers/playlist.controller.js
@@ -112,9 +112,10 @@ const playlistController = {
   update: async (req, res, next) => {
     try {
       let playlistData;
+      const uid = req.user.id;
       const playlistId = req.params.id;
 
-      const { title, photo, status, sound, favouriteValue } = req.body;
+      const { title, photo, status, sounds, favouriteValue } = req.body;
 
       if (title && (title.length < 3 || title.length > 100)) {
         return next(new ApiError(400, "Thông tin không hợp lệ!"));
@@ -122,33 +123,45 @@ const playlistController = {
 
       const updateData = { title, status };
 
-      if (photo && photo.split(":")[0] == "data") {
-        playlistData = await playlistModel.findById(playlistId);
+      const playlist = await playlistModel.find({ user: uid, _id: playlistId });
+      if (playlist) {
+        if (photo && photo.split(":")[0] == "data") {
+          playlistData = await playlistModel.findById(playlistId);
 
-        if (!playlistData) {
-          return next(new ApiError(404, "Không tìm thấy playlist!"));
+          if (!playlistData) {
+            return next(new ApiError(404, "Không tìm thấy playlist!"));
+          }
+
+          const fileName = Date.now();
+          const photoPath =
+            "uploads/images/" + fileNameGeneral(fileName, typeFile(photo));
+          fs.writeFileSync(photoPath, photo.split(",")[1], "base64");
+          updateData.photo = photoPath;
+
+          if (playlistData.photo) deleteFile(playlistData.photo);
         }
-
-        const fileName = Date.now();
-        const photoPath =
-          "uploads/images/" + fileNameGeneral(fileName, typeFile(photo));
-        fs.writeFileSync(photoPath, photo.split(",")[1], "base64");
-        updateData.photo = photoPath;
-
-        if (playlistData.photo) deleteFile(playlistData.photo);
-      }
-
-      await performCRUD(playlistModel, "update", {
-        id: playlistId,
-        ...updateData,
-      });
-
-      if (sound) {
-        console.log(sound);
         await performCRUD(playlistModel, "update", {
           id: playlistId,
-          $push: { sounds: sound },
+          ...updateData,
         });
+
+        if (sounds && sounds.length > 0) {
+          const existingPlaylist = await playlistModel.findById(playlistId);
+          const existingSounds = existingPlaylist
+            ? existingPlaylist.sounds
+            : [];
+
+          const newSounds = sounds.filter(
+            (sound) => !existingSounds.includes(sound)
+          );
+
+          if (newSounds.length > 0) {
+            await performCRUD(playlistModel, "update", {
+              id: playlistId,
+              $push: { sounds: { $each: newSounds } },
+            });
+          }
+        }
       }
 
       if (favouriteValue) {
@@ -185,6 +198,48 @@ const playlistController = {
       return res
         .status(204)
         .json({ statusCode: 204, message: "Xóa thành công!" });
+    } catch (error) {
+      console.log(error);
+      return next(new ApiError(500, "Đã xảy ra lỗi. Vui lòng thử lại sau."));
+    }
+  },
+  getPlaylistByOther: async (req, res, next) => {
+    try {
+      let modifiedPlaylistData;
+      const { token, keyword, limit = 1000, page = 1 } = req.query;
+      const startIndex = (page - 1) * limit;
+
+      const playlist = await playlistModel
+        .find({ user: keyword })
+        .populate({
+          path: "user",
+          select:
+            "-username -password -createdAt -updatedAt -sounds -purchases",
+        })
+        .skip(startIndex)
+        .limit(limit)
+        .sort({ createdAt: -1 });
+
+      const totalResults = await playlistModel.countDocuments({
+        user: keyword,
+      });
+
+      if (token) {
+        const uid = verifyAccessToken(token)?.id;
+        if (uid) modifiedPlaylistData = playlist;
+      } else {
+        modifiedPlaylistData = playlist.filter(
+          (item) => item.status != "private"
+        );
+      }
+
+      return res.status(200).json({
+        statusCode: 200,
+        data: modifiedPlaylistData,
+        page,
+        limit,
+        total: totalResults,
+      });
     } catch (error) {
       console.log(error);
       return next(new ApiError(500, "Đã xảy ra lỗi. Vui lòng thử lại sau."));
